+++ src/card/sets/y2004/mod.rs
@@ -1,0 +1,5 @@
+//! Card sets first released in 2004.
+
+pub(super) mod darksteel;
+
+pub(super) use super::{CardRecord, PrintingRecord};